+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Bug Fixes
 - Dark mode now works without delay
+- Fixed timestamps on Subject Details Popup ([#16](https://github.com/digas99/wanikani-kanji-highlighter/issues/16))
 
 ## Popup
 - Improved interactions with Future Reviews chart